backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
mysqlclient==2.2.0
pydantic==2.5.2
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.2 